Switch license to GPL2 (drama)
As alluded to in the comment, it's unclear what the license of these
backends _should_ be. libmysqlclient is GPL2, but the connector code for
libgit2 is written to be GPL2 with linking exception -- which might be
invalid (who can say).

Given that pygit2 is GPL2, and we link (broadly) with pygit2, uprate the
license to the strongest compatible one, GPL2.